Moved and renamed UserDAO to UserDTO, implemented CORS configuration for front end configuration
