+++ src/main/java/com/mtiteiu/clinic/controllers/AuthenticationController.java
@@ -4,6 +4,7 @@ import com.mtiteiu.clinic.security.AuthRequest;
 import com.mtiteiu.clinic.security.JwtUtils;
 import com.mtiteiu.clinic.service.UserService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/auth")
 @RequiredArgsConstructor
+@Slf4j
 public class AuthenticationController {
 
     private final UserService userDetailsService;
@@ -25,12 +27,17 @@ public class AuthenticationController {
     @PostMapping("/authenticate")
     public ResponseEntity<String> authenticate(@RequestBody AuthRequest authRequest) {
 
-        authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequest.getEmail(), authRequest.getPassword()));
-
-        final UserDetails user = userDetailsService.loadUserByUsername(authRequest.getEmail());
-        if (user != null) {
-            return ResponseEntity.ok(jwtUtils.generateToken(user));
+        ResponseEntity<String> response = null;
+        try{
+            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authRequest.getEmail(), authRequest.getPassword()));
+            final UserDetails user = userDetailsService.loadUserByUsername(authRequest.getEmail());
+            if (user != null) {
+                return ResponseEntity.ok(jwtUtils.generateToken(user));
+            }
+        } catch (Exception ex){
+            log.error("Error while trying to authenticate user! {}", ex.getMessage());
+            response = ResponseEntity.status(400).body(String.format("Some error has occurred while trying to authenticate the user! %s", ex.getMessage()));
         }
-        return ResponseEntity.status(400).body("Some error has occurred");
+        return response;
     }
 }
