+++ src/test/java/com/mtiteiu/clinic/TestUtils.java
@@ -1,6 +1,6 @@
 package com.mtiteiu.clinic;
 
-import com.mtiteiu.clinic.dao.UserDTO;
+import com.mtiteiu.clinic.dto.UserDTO;
 import com.mtiteiu.clinic.model.patient.PatientDetails;
 import com.mtiteiu.clinic.model.user.Role;
 import com.mtiteiu.clinic.model.user.User;
